Prevent start.sh getting stuck in a loop
In some conditions the trap handler would run in response to the kill,
which itself was invoked from the trap handler.  Madness ensues.

+++ scripts/start.sh
@@ -59,9 +59,14 @@ while [ "$#" -gt 0 ]; do
     shift
 done
 
+TRAPPED=0
+
 function stop_solr() {
-    solr_dist/bin/solr stop -p $solr_port
-    kill 0
+    if [ "$TRAPPED" = "0" ]; then
+        TRAPPED=1
+        solr_dist/bin/solr stop -p $solr_port
+        kill 0
+    fi
 }
 
 function fail() {
